refactor(mergeAll): move support types under mergeAll.ts
